Use timers for triggering sync instead of invocations
Previously we would trigger a sync for every change to the document.
This meant that we would never batch changes up for sending, only
testing that single change sets get synced correctly.

Since timers automatically try to jump in at every opportunity they
interleave the operations for us, so we use that to trigger syncing.
This also updates the sync code to generate and work with batches of
changes. Sync messages are unnafected in the logic but should get more
interesting cases now. Save load syncing is unchanged.