Debug synced bytes as base64 and optimise sync for changes
Base64 makes the byte vectors more readable in the explorer whilst still
making them usable for copying to a test case.

As for optimising sync: previously generating the changes required a
mutable document, which would cause another transition in stateright so
we add a check before that to cut down on clones.